+++ src/test/framework/behavior.lua
@@ -13,13 +13,13 @@ function patrol:run()
         self._cur_dis_ = 0
         self._dis_ = Rand(50, 100)
         self._stage_ = 2
-        print("patrol has set the distance.")
+        print("patrol has set the distance.", os.clock())
     elseif self._cur_dis_ >= self._dis_ then
-        print("patrol is finished.")
+        print("patrol is finished.", os.clock())
         return self:success()
     else
         self._cur_dis_ = self._cur_dis_ + Rand(10, 15)
-        print(table.concat{self.tree_:getParam("目標"), " is patroling."})
+        print(table.concat{self.tree_:getParam("目標"), " is patroling."}, os.clock())
     end
 
     self:running()
@@ -28,10 +28,10 @@ end
 local b = Node("碰到敵人")
 function b:run()
     if Rand(1, 100) < 51 then
-        print("encounter an enemy.")
+        print("encounter an enemy.", os.clock())
         self:success()
     else
-        print("It's safe.")
+        print("It's safe.", os.clock())
         self:fail()
     end
 end
@@ -40,10 +40,10 @@ local c = Node("生命值檢測")
 
 function c:run()
     if Rand(1, 100) < self._args_[1] then
-        print("可以攻擊")
+        print("可以攻擊", os.clock())
         self:success()
     else
-        print("它太強了")
+        print("它太強了", os.clock())
         self:fail()
     end
 end
@@ -68,46 +68,48 @@ function escape:run()
     self._cur_dis_ = self._cur_dis_ + Rand(20, 50)
 
     if self._cur_dis_ > self._dis_ then
-        print(self.tree_:getParam("目標"), "has escaped.")
-        self:suceess()
+        print(self.tree_:getParam("目標"), "has escaped.", os.clock())
+        self:success()
     else
-        print(self.tree_:getParam("目標"), "is escaping.")
+        print(self.tree_:getParam("目標"), "is escaping.", os.clock())
         self:running()
     end
 end
 
 local death = Node("死亡")
 function death:run()
-    if Rand(1, 10) < 5 then
+    if Rand(1, 100) < 30 then
+        print("death", os.clock())
         self:success()
     else
+        print("alive", os.clock())
         self:fail()
     end
 end
 
 local attack = Node("攻擊")
 function attack:run()
-    print(self.tree_:getParam("目標"), "is attacking.")
+    print(self.tree_:getParam("目標"), "is attacking.", os.clock())
     self:running()
 end
 
-local bt = BT:new(123, {id="Condition", nodes={
-        {id="碰到敵人"},
-        {id="Condition", nodes={
-            {id="生命值檢測", args={30}},
-            {id="Condition", nodes={
-                {id="距離檢測", args={100}},
-                {id="None"},
-                {id="逃跑"}
-            }},
-            {id="Condition", nodes={
-                {id="死亡"},
-                {id="None"},
-                {id="攻擊"}
-            }}
-        }},
-        {id="巡邏"}
-    }}):setParam("目標", "步兵123"):run()
+-- local bt = BT:new(123, {id="Condition", nodes={
+--         {id="碰到敵人"},
+--         {id="Condition", nodes={
+--             {id="生命值檢測", args={30}},
+--             {id="Condition", nodes={
+--                 {id="距離檢測", args={100}},
+--                 {id="None"},
+--                 {id="逃跑"}
+--             }},
+--             {id="Condition", nodes={
+--                 {id="死亡"},
+                -- {id="None"},
+                -- {id="攻擊"}
+--             }}
+--         }},
+--         {id="巡邏"}
+--     }}):setParam("目標", "步兵123"):run()
 
 -- print(bt)
 -- bt:insert("2", Node("攻擊"):new())
@@ -127,7 +129,18 @@ local bt = BT:new(123, {id="Condition", nodes={
 --     }},
 --     {id="等待", args={1}}
 -- }}):setParam("目標", "工作人員"):run()
-print("a", os.clock())
-print(BT:new(1, {id="等待", args={1}}):run())
-print("b", os.clock())
+
+-- print("a", os.clock())
+-- print(BT:new(1, {id="等待", args={1}}):run())
+-- print("b", os.clock())
+
+print(BT:new(1, {id="Timer", args={0.5}, nodes={
+    {id="逃跑"},
+    {id="Condition", nodes={
+        {id="死亡"},
+        {id="None"},
+        {id="攻擊"}
+    }},
+    {id="巡邏"}
+}}):setParam("目標", "工作人員"):run())